+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityEditModel.java
@@ -114,7 +114,7 @@ public class SwingEntityEditModel extends AbstractEntityEditModel {
 	public final void clearComboBoxModels() {
 		synchronized (comboBoxModels) {
 			for (FilterComboBoxModel<?> comboBoxModel : comboBoxModels.values()) {
-				comboBoxModel.clear();
+				comboBoxModel.items().clear();
 			}
 		}
 	}