+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/component/EntityComboBoxModel.java
@@ -342,11 +342,6 @@ public final class EntityComboBoxModel implements FilterComboBoxModel<Entity> {
 	}
 
 	@Override
-	public void clear() {
-		comboBoxModel.clear();
-	}
-
-	@Override
 	public ComboBoxSelection<Entity> selection() {
 		return comboBoxModel.selection();
 	}