+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntitySearchField.java
@@ -898,7 +898,7 @@ public final class EntitySearchField extends HintTextField {
 							.okAction(selectControl)
 							.show();
 
-			table.model().clear();
+			table.model().items().clear();
 			table.searchField().setText("");
 		}